==> Step 0 Clean NULL bytes from EMu CSV files.ipynb <==
import pandas as pd

test = pd.read_csv('raw_data/Taxonomy/AutBasio.csv')
test.head()

with open('raw_data/Taxonomy/AutBasio_clean.csv','w') as outfile:
    with open('raw_data/Taxonomy/AutBasio.csv','r') as infile:
        for line in infile:
            line = line.replace('\x00', '')
            outfile.write(line)

test = pd.read_csv('raw_data/Taxonomy/AutBasio_clean.csv')
test.head()

import os

csv_list = []
for root, dirs, files in os.walk('.'):
    for file in files:
        if file.endswith('.csv'):
            csv_file = os.path.join(root, file)
            csv_list.append(csv_file)
            print(csv_file)

for csv_file in csv_list:
    with open(csv_file.replace('.csv', '_clean.csv'),'w', encoding='utf-8') as outfile:
        with open(csv_file,'r', encoding='latin1') as infile:
            for line in infile:
                line = line.replace('\x00', '').strip()
                if len(line):
                    outfile.write(line + '\n')
