==> Step 1 Comparing BOLD taxonomy against EMu.ipynb <==
import pandas as pd
import numpy as np

## Read in BOLD spreadsheet

bold_file = 'raw_data/asilo_bold.xlsx'

bold_df = pd.read_excel(bold_file, sheet_name='Lab Sheet', skiprows=2)
drop_cols = ['Field ID','Catalog Num','Collection Date','Identification',
             'Life Stage','Extra Info','Voucher Type','Institution','Notes']
print(bold_df.columns)
keep_cols = [col for col in bold_df.columns if col not in drop_cols]
bold_df = bold_df[keep_cols]
bold_df.info()

voucher_df = pd.read_excel(bold_file, sheet_name='Voucher Info', skiprows=2)
voucher_df.info()

tax_df = pd.read_excel(bold_file, sheet_name='Taxonomy', skiprows=2)
tax_df.info()

specimen_df = pd.read_excel(bold_file, sheet_name='Specimen Details', skiprows=2)
specimen_df.info()

collection_df = pd.read_excel(bold_file, sheet_name='Collection Data', skiprows=2)
collection_df.info()

for df in [voucher_df, tax_df, specimen_df, collection_df]:
    bold_df = bold_df.merge(df, on='Sample ID')
bold_df.info()

bold_df = bold_df.dropna(axis=1, how='all')
bold_df.info()

bold_df.to_csv('processed_data/asilo_bold.tsv', index=False, sep='\t')

## Pull out taxonomic names, and clean up

bold_df = pd.read_csv('processed_data/asilo_bold.tsv', sep='\t')
bold_df.info()

tax_cols = ['Phylum','Class','Order','Family','Subfamily','Genus','Species']
bold_tax = bold_df.groupby(tax_cols).size().reset_index(name='tax_count')
bold_tax.head()

len(bold_tax)

bold_tax = bold_tax.rename(columns={'Species':'Scientific Name'})

bold_tax['Species'] = bold_tax['Scientific Name'].str.split(' ').str.get(1)
bold_tax['Species'] = bold_tax['Species'].replace({'sp.':np.nan,
                                                 'spp.':np.nan})
bold_tax['Species'] = bold_tax['Species'].fillna('unidentified')
bold_tax[['Genus','Scientific Name','Species']].head()

len(bold_tax[bold_tax['Species'] == 'unidentified'])

def bold_name(row):
    if pd.notnull('Genus'):
        return "{} {}".format(row['Genus'], row['Species'])
    else:
        return "{} {}".format(row['Family'], row['Species'])

bold_tax['BOLD Name'] = bold_tax.apply(bold_name, axis=1)
bold_tax.head()

## Read in EMu taxonomy file

emu_tax_base = pd.read_csv('raw_data/Taxonomy/etaxonom_clean.csv')
emu_tax_base.info()

len(emu_tax_base[emu_tax_base['ClaSpecies'] == 'unidentified'])

## Match up species with EMu taxonomy

merged = emu_tax_base.merge(bold_tax, left_on=['ClaFamily','ClaGenus','ClaSpecies'],
                            right_on=['Family','Genus','Species'])
len(merged)

merged.head()

## Export matched species, and new species to upload

matched_names = merged['BOLD Name'].tolist()
not_matched = bold_tax[~bold_tax['BOLD Name'].isin(matched_names)]
len(not_matched)

not_matched.to_csv('processed_data/taxa_unmatched.tsv', sep='\t', index=False)

print(len(not_matched[not_matched['Species'] != 'unidentified']))
print(len(not_matched[not_matched['Species'] == 'unidentified']))

matched_cols = list(bold_tax.columns) + ['irn']
matched_tax = merged[matched_cols]
matched_tax.head()

matched_tax.to_csv('processed_data/taxa_matched.tsv', sep='\t', index=False)

len(matched_tax[matched_tax['Species'] != 'unidentified'])

len(matched_tax[matched_tax['Species'] == 'unidentified'])
