==> Step 2 Finding Authors for New Taxon.ipynb <==
import pandas as pd
import numpy as np

## Read in new species to upload

unmatched_taxa = pd.read_csv('processed_data/taxa_unmatched.tsv', sep='\t')
unmatched_taxa.head()

taxa_to_search = unmatched_taxa[unmatched_taxa['Species'] != 'unidentified']
len(taxa_to_search)

## Use GBIF webservice to find authors for these new species

import requests
from unidecode import unidecode

def gbif_author_search(name):
    search_url = 'http://api.gbif.org/v1/species/match'
    params = {'kingdom':'Animalia',
              'name':name,
              'verbose':'true'}
    r = requests.get(search_url, params)
    search_results = r.json()
    confidence = search_results['confidence']
    if (int(confidence) > 90) and (search_results['matchType'] == 'EXACT'):
        gbif_id = search_results['usageKey']
        fetch_url = 'http://api.gbif.org/v1/species/{}'.format(gbif_id)
        r = requests.get(fetch_url)
        fetch_results = r.json()
        authorship = unidecode(fetch_results['authorship'])
        return authorship
    return np.nan

taxa_to_search['gbif_authorship'] = taxa_to_search['BOLD Name'].apply(gbif_author_search)

taxa_to_search.head()

len(taxa_to_search[pd.isnull(taxa_to_search['gbif_authorship'])])

taxa_to_search[pd.isnull(taxa_to_search['gbif_authorship'])].head()

## Export new author names

authors_found = taxa_to_search[pd.notnull(taxa_to_search['gbif_authorship'])]
len(authors_found)

authors_found.to_csv('processed_data/taxa_unmatched_authors_found.tsv',
                     index = False, sep='\t')

## Export species whose authors could not be found

authors_lost = taxa_to_search[pd.isnull(taxa_to_search['gbif_authorship'])]
len(authors_lost)

authors_lost.head()

authors_lost.to_csv('processed_data/taxa_unmatched_authors_lost.tsv',
                     index = False, sep='\t')
