==> Step 3 Generating Parties upload for new authors and collectors.ipynb <==
import pandas as pd
import numpy as np

## Read in EMu parties dump

parties_core = pd.read_csv('raw_data/Parties/eparties_clean.csv')
parties_core.info()

parties_core.head()

roles_df = pd.read_csv('raw_data/Parties/NamRoles_clean.csv')
print(len(roles_df))
roles_df.head()

roles_df['NamRoles'].value_counts()

roles_df['NamRoles'] = roles_df['NamRoles'].replace({'collector':'Collector',
                                                      'author':'Author',
                                                      'identifier':'Identifier'})
roles_df['NamRoles'].value_counts()

party_roles = parties_core.merge(roles_df, on='eparties_key', how='left')

keep_cols = ['irn','NamFullName','NamFirst','NamLast','NamRoles']
party_roles = party_roles[keep_cols]
party_roles.info()

party_roles = party_roles.fillna('')

squished = party_roles.groupby(['irn','NamFullName','NamFirst','NamLast'])['NamRoles'].agg(lambda role: '; '.join(role)).reset_index()
squished.info()

emu_authors = squished[squished['NamRoles'].str.contains('Author')]
print(len(emu_authors))
emu_authors.head()

basio_df = pd.read_csv('raw_data/Taxonomy/AutBasio_clean.csv')
comb_df = pd.read_csv('raw_data/Taxonomy/AutCombA_clean.csv')

author_irns = pd.concat([basio_df['irn'], comb_df['irn']])
author_irns = author_irns.dropna().astype('int')
author_irns.head()

author_irn_counts = author_irns.value_counts().rename_axis('irn').reset_index(name='Author Count')
author_irn_counts.head()

emu_author_counts = emu_authors.merge(author_irn_counts, how='left')
emu_author_counts.head()

emu_author_counts = emu_author_counts.sort_values('Author Count', ascending=False)
emu_author_counts.head()

len(emu_author_counts)

top_emu_authors = emu_author_counts.drop_duplicates(subset=['NamFullName'], keep='first')
len(top_emu_authors)

## Read in taxonomic authors list

tax_authors = pd.read_csv('processed_data/taxa_unmatched_authors_found.tsv', sep='\t')
tax_authors.head()

def author_split(row):
    if pd.notnull(row['gbif_authorship']):
        author_line = str(row['gbif_authorship']).strip()
        if author_line.startswith('('):
            row['author_type'] = 'basionym'
            author_line = author_line.lstrip('(').rstrip(')').strip()
        else:
            row['author_type'] = 'regular'
        authors, year = author_line.split(', ')
        authors_split = authors.split(' & ')
        if row['author_type'] == 'regular':
            author_base = 'AutCombAuthorsRef_tab'
            year_base = 'AutCombYear'
        else:
            author_base = 'AutBasionymAuthorsRef_tab'
            year_base = 'AutBasionymYear'
        for idx, author in enumerate(authors_split):
            author_idx = '{}.{}'.format(author_base, idx + 1)
            row[author_idx] = author
        row[year_base] = year
    return row

tax_authors = tax_authors.apply(author_split, axis=1)
tax_authors.head(10)

all_authors = tax_authors['AutBasionymAuthorsRef_tab.1'].dropna().tolist() + \
              tax_authors['AutBasionymAuthorsRef_tab.2'].dropna().tolist() + \
              tax_authors['AutCombAuthorsRef_tab.1'].dropna().tolist() + \
              tax_authors['AutCombAuthorsRef_tab.2'].dropna().tolist()
print(len(all_authors))
print(len(set(all_authors)))

from collections import Counter

author_counter = Counter(all_authors)
author_counts = pd.DataFrame.from_dict(author_counter, orient='index').reset_index()
author_counts.columns = ['Author','Dataset Count']
author_counts = author_counts.sort_values('Dataset Count',ascending=False)
author_counts.head()

## Match up authors with EMu IRNs

matched_authors = author_counts.merge(top_emu_authors, left_on='Author', right_on='NamFullName',
                              how='left')
matched_authors.head()

print(len(matched_authors[pd.isnull(matched_authors['irn'])]))
print(len(matched_authors[pd.notnull(matched_authors['irn'])]))

unmatched_authors = matched_authors[pd.isnull(matched_authors['irn'])][['Author','Dataset Count']]
unmatched_authors.head()

unmatched_authors['NamRoles_tab(1)'] = 'Author'
unmatched_authors = unmatched_authors.drop('Dataset Count', axis=1)

unmatched_authors = unmatched_authors.rename(columns={'Author':'NamBriefName'})
unmatched_authors['NamLast'] = unmatched_authors['NamBriefName']
unmatched_authors.head()


unmatched_authors.to_csv('upload_files/parties_tax_authors.csv', index=False)

matched_authors_irns = matched_authors[pd.notnull(matched_authors['irn'])]
matched_authors_irns.to_csv('processed_data/authors_matched_irns.tsv', sep='\t', index=False)

## Read in BOLD spreadsheet, and pull out collectors and identifiers

bold_df = pd.read_csv('processed_data/asilo_bold.tsv', sep='\t')

bold_collectors = bold_df['Collectors'].str.split(r',\s*', expand=True).rename(columns = lambda x: "Collector"+str(x+1))
bold_collectors.head()

Bring back in BOLD Process ID, so we can map back IRNs later.

bold_collectors = bold_collectors.merge(bold_df[['Process ID','Collectors']], 
                                                left_index=True, right_index=True)
bold_collectors.head()

collector_columns = ['Collector1','Collector2','Collector3','Collector4','Collector5']
for col in collector_columns:
    bold_collectors[col] = bold_collectors_process[col].str.replace('.','. ')
    bold_collectors[col] = bold_collectors_process[col].str.rstrip('.')
    bold_collectors[col] = bold_collectors_process[col].str.rstrip()
bold_collectors.head()

bold_collectors.to_csv('processed_data/collectors_bold_mapback.tsv', 
                               index=False, sep='\t')

Export to file for extra manual formatting, and then read back in.

bold_collectors = pd.read_csv('processed_data/collectors_bold_mapback_fixed.tsv', sep='\t')

all_collectors = bold_collectors['Collector1'].dropna().tolist() + \
                 bold_collectors['Collector2'].dropna().tolist() + \
                 bold_collectors['Collector3'].dropna().tolist() + \
                 bold_collectors['Collector4'].dropna().tolist() + \
                 bold_collectors['Collector5'].dropna().tolist()
print(len(all_collectors))

collector_counter = Counter(all_collectors)
collector_counts = pd.DataFrame.from_dict(collector_counter, orient='index').reset_index()
collector_counts.columns = ['Collector','Dataset Count']
collector_counts = collector_counts.sort_values('Dataset Count',ascending=False)
collector_counts.head()

def split_collector_names(row):
    name_split = str(row['Collector']).split(' ')
    row['NamLast'] = name_split[-1]
    if len(name_split) > 1:
        row['NamFirst'] = name_split[0]
    if len(name_split) > 2:
        row['NamMiddle'] = ' '.join(name_split[1:-1])
    return row
collector_counts = collector_counts.apply(split_collector_names, axis=1)
collector_counts.head()

collector_counts = collector_counts.sort_values('NamLast')
collector_counts.head()

collector_counts.to_csv('processed_data/collector_counts.tsv', index=False, sep='\t')

len(collector_counts)

emu_collectors = squished[squished['NamRoles'].str.contains('Collector')]
emu_collectors = emu_collectors.sort_values('irn', ascending=True)
emu_collectors = emu_collectors.drop_duplicates(subset=['NamFullName'], keep='first')
len(emu_collectors)

collectors_merged = collector_counts.merge(emu_collectors, 
                                            left_on='Collector', right_on='NamFullName',
                                            how='left')
collectors_merged.head()

collectors_matched = collectors_merged[pd.notnull(collectors_merged['irn'])]
print(len(collectors_matched))
collectors_unmatched = collectors_merged[pd.isnull(collectors_merged['irn'])]
print(len(collectors_unmatched))
collectors_matched.head()

collectors_found = collectors_matched['Collector'].tolist()
print(collectors_found[:5])
collectors_to_upload = collector_counts[~collector_counts['Collector'].isin(collectors_found)]
collectors_to_upload.head()

collectors_to_upload = collectors_to_upload.rename(columns={'Collector':'NamBriefName'})
collectors_to_upload['NamRoles_tab(1)'] = 'Collector'
print(len(collectors_to_upload))
collectors_to_upload.head()

## Pull out Identifiers and Match up with EMu

bold_iders = bold_df['Identifier'].str.split(', ', expand=True).rename(columns = lambda x: "Identifier"+str(x+1))
bold_iders.head()

all_iders = bold_iders['Identifier1'].dropna().tolist() + \
            bold_iders['Identifier2'].dropna().tolist()
print(len(all_iders))

id_counter = Counter(all_iders)
id_counts = pd.DataFrame.from_dict(id_counter, orient='index').reset_index()
id_counts.columns = ['Identifier','Dataset Count']
id_counts = id_counts.sort_values('Dataset Count',ascending=False)
id_counts.head()

Drop "USNM Curators" as an identifier. It was used as a default value in BOLD, and isn't a valid entry in EMu.

id_counts = id_counts.drop(0)
id_counts.head()

len(id_counts)

id_counts.to_csv('processed_data/identifier_counts.tsv', sep='\t', index=False)

## Match up collectors and identifiers

emu_identifiers = squished[squished['NamRoles'].str.contains('Identifier')]
print(len(emu_identifiers))
emu_identifiers = emu_identifiers.sort_values('irn', ascending=True)
emu_identifiers = emu_identifiers.drop_duplicates(subset=['NamFullName'], keep='first')
print(len(emu_identifiers))

identifiers_merged = id_counts.merge(emu_identifiers, 
                                            left_on='Identifier', right_on='NamFullName',
                                            how='left')
identifiers_matched = identifiers_merged[pd.notnull(identifiers_merged['irn'])]
print(len(identifiers_matched))
identifiers_unmatched = identifiers_merged[pd.isnull(identifiers_merged['irn'])]
print(len(identifiers_unmatched))
identifiers_matched.head()

identifiers_found = identifiers_matched['Identifier'].tolist()
iders_to_upload = id_counts[~id_counts['Identifier'].isin(identifiers_found)]
iders_to_upload.head()

def split_identifier_names(row):
    name_split = str(row['Identifier']).split(' ')
    row['NamLast'] = name_split[-1]
    if len(name_split) > 1:
        row['NamFirst'] = name_split[0]
    if len(name_split) > 2:
        row['NamMiddle'] = ' '.join(name_split[1:-1])
    return row
iders_to_upload = iders_to_upload.apply(split_identifier_names, axis=1)
iders_to_upload.head()

iders_to_upload = iders_to_upload.rename(columns={'Identifier':'NamBriefName'})
iders_to_upload['NamRoles_tab(1)'] = 'Identifier'
print(len(iders_to_upload))
iders_to_upload.head()

## Export new parties to be uploaded

parties_to_upload = pd.concat([iders_to_upload, collectors_to_upload,unmatched_authors], sort=True)
print(len(parties_to_upload))
parties_to_upload.head()

parties_to_upload = parties_to_upload.drop(columns=['Dataset Count'])
parties_to_upload.head()

parties_to_upload = parties_to_upload.sort_values('NamLast')
parties_to_upload

parties_to_upload.to_csv('upload_files/parties_to_upload.csv', index=False)

## Export parties IRN matches
